Module Dns_resolver_mirage.Make

Parameters

module R : Mirage_random.S

module P : Mirage_clock.PCLOCK

module M : Mirage_clock.MCLOCK

module T : Mirage_time.S

module S : Mirage_stack.V4V6

Signature

val resolver : S.t -> ?root:bool -> ?timer:int -> ?udp:bool -> ?tcp:bool -> ?tls:Tls.Config.server -> ?port:int -> ?tls_port:int -> Dns_resolver.t -> unit

resolver stack ~root ~timer ~udp ~tcp ~tls ~port ~tls_port resolver registers a caching resolver on the provided protocols udp, tcp, tls using port for udp and tcp (defaults to 53), tls_port for tls (defaults to 853) using the resolver configuration. The timer is in milliseconds and defaults to 500 milliseconds.