Module Dns_resolver_utils

val scrub : ?⁠mode:[ `Recursive | `Stub ] -> [ `raw ] Domain_name.t -> Dns.Packet.Question.qtype -> Dns.Packet.t -> ((Dns.Rr_map.k * [ `raw ] Domain_name.t * Dns_resolver_cache.rank * Dns_resolver_cache.res) list, Dns.Rcode.t) Stdlib.result

scrub ~mode bailiwick packet returns a list of entries to-be-added to the cache. This respects only in-bailiwick resources records, and qualifies the packet. The purpose is to avoid cache poisoning by not accepting all resource records.

val invalid_soa : [ `raw ] Domain_name.t -> Dns.Soa.t

invalid_soa name returns a stub SOA for name.