Module Dns_resolver_mirage.Make

Parameters

R : Mirage_random.S

P : Mirage_clock.PCLOCK

M : Mirage_clock.MCLOCK

T : Mirage_time.S

S : Mirage_stack.V4

Signature

val resolver : S.t -> ?⁠root:bool -> ?⁠timer:int -> ?⁠port:int -> Dns_resolver.t -> unit

resolver stack ~root ~timer ~port resolver registers a caching resolver on the provided port (both udp and tcp) using the resolver configuration. The timer is in milliseconds and defaults to 500 milliseconds.